Move items around and fix serde error
